chore: rename to useReduxData
