+++ src/components/DownloadButton.tsx
@@ -1,4 +1,4 @@
-import { useCanvasData, useCanvasImageData } from "./lib/hooks/useReduxHooks"
+import { useCanvasData, useCanvasImageData } from "./lib/hooks/useReduxData"
 import { useRef } from "react"
 
 import DownloadIcon from "./icons/DownloadIcon"
